Move backend-specific relations to their own directories
