+++ symbolic_pymc/relations/theano/linalg.py
@@ -11,8 +11,8 @@ from kanren.core import lall, lallgreedy
 from kanren.goals import not_equalo, conso
 from kanren.term import term, operator, arguments
 
-from ..theano.meta import mt
-from ..unify import etuple, tuple_expression, ExpressionTuple
+from ...theano.meta import mt
+from ...unify import etuple, tuple_expression, ExpressionTuple
 
 
 mt.nlinalg.qr_full = mt(QRFull("reduced"))
