+++ symbolic_pymc/theano/unify.py
@@ -38,7 +38,6 @@ arguments.add((tt.Variable,), lambda x: arguments(metatize(x)))
 
 term.add((tt.Op, ExpressionTuple), lambda op, args: term(metatize(op), args))
 
-tuple_expression.add(tt_class_abstractions,
-                     lambda x: tuple_expression(metatize(x)))
+tuple_expression.add(tt_class_abstractions, lambda x: tuple_expression(metatize(x)))
 
 __all__ = []
