+++ tests/theano/test_conjugates.py
@@ -8,7 +8,7 @@ from theano.gof.graph import inputs as tt_inputs
 from symbolic_pymc.theano.random_variables import MvNormalRV, observed
 from symbolic_pymc.theano.opt import KanrenRelationSub, FunctionGraph
 from symbolic_pymc.theano.utils import optimize_graph
-from symbolic_pymc.relations.conjugates import conjugate_posteriors
+from symbolic_pymc.relations.theano.conjugates import conjugate_posteriors
 
 
 @pytest.mark.usefixtures("run_with_theano")
