+++ tests/theano/test_linalg.py
@@ -11,8 +11,8 @@ from symbolic_pymc.theano.random_variables import NormalRV, observed
 from symbolic_pymc.theano.meta import mt
 from symbolic_pymc.theano.opt import eval_and_reify_meta, FunctionGraph
 from symbolic_pymc.unify import (etuple, tuple_expression)
-from symbolic_pymc.relations.linalg import (normal_normal_regression, buildo,
-                                            normal_qr_transform)
+from symbolic_pymc.relations.theano.linalg import (normal_normal_regression, buildo,
+                                                   normal_qr_transform)
 
 from kanren import run, eq, var
